Update graphexp.html
Makes info boxes selectable to allow copying data. By necessity you can only interact with the visualization in the empty space between the left and right boxes (if displayed) but this is much more intuitive than wholly blocking interaction with the text. Note, if this is a major problem, setting some max width formatting on the info boxes would be good.