Add possibility to fix node position from the DB (#24)
Change to graphioGremlin.js extract_infov3 so that if node properties "graphexpx" or "graphexpy" are found in Gremlin DB (they should have numerical values) then these are used to constrain the x or y coordinates of the nodes in the visualisation. Any dimension which is not constrained, as well as any nodes which are not constrained in either dimension, are subject to the force simulation. Thus, adding only "graphexpy" properties can be used to display hierarchies, whether local or global, that may exist in your data. I recommend multiplying each desired hierarchical level by 50-100 in the first case in order to work well with the default graphexp visualisation settings. (#24)